requirements.txt
flask
flask-cors
supabase
python-dotenv
langchain
langchain-community
langchain-google-genai
gunicorn
faiss-cpu
sentence-transformers